# [stdlib] Greatly optimize unicode casing