--- a/mojo/stdlib/benchmarks/collections/bench_string.mojo
+++ b/mojo/stdlib/benchmarks/collections/bench_string.mojo
@@ -173,8 +173,9 @@ def bench_string_lower[
 
     @always_inline
     def call_fn() unified {read}:
-        var res = black_box(items).lower()
-        keep(res)
+        for _ in range(10**6 // length):
+            var res = black_box(items).lower()
+            keep(res)
 
     b.iter(call_fn)
 
@@ -190,8 +191,9 @@ def bench_string_upper[
 
     @always_inline
     def call_fn() unified {read}:
-        var res = black_box(items).upper()
-        keep(res)
+        for _ in range(10**6 // length):
+            var res = black_box(items).upper()
+            keep(res)
 
     b.iter(call_fn)