Greatly optimize unicode casing by adding fast paths for ASCII and Latin-1 encodings.

## Benchmark results

CPU: Intel® Core™ i7-7700HQ

|bench | old value (ms) | new value (ms) | markdown perc | magnitude|
|-- | -- | -- | -- | --|
|bench_string_lower[10] | 72.42 | 15.50 | 0.79 | 4.67|
|bench_string_upper[10] | 121.40 | 22.11 | 0.82 | 5.49|
|bench_string_lower[30] | 74.95 | 15.39 | 0.79 | 4.87|
|bench_string_upper[30] | 125.36 | 21.35 | 0.83 | 5.87|
|bench_string_lower[50] | 77.73 | 14.76 | 0.81 | 5.27|
|bench_string_upper[50] | 132.22 | 20.05 | 0.85 | 6.59|
|bench_string_lower[100] | 86.46 | 14.48 | 0.83 | 5.97|
|bench_string_upper[100] | 119.76 | 19.96 | 0.83 | 6.00|
|bench_string_lower[1000] | 69.18 | 14.03 | 0.80 | 4.93|
|bench_string_upper[1000] | 125.47 | 19.48 | 0.84 | 6.44|
|bench_string_lower[10000] | 67.78 | 13.90 | 0.79 | 4.88|
|bench_string_upper[10000] | 125.44 | 19.44 | 0.85 | 6.45|
|bench_string_lower[100000] | 67.40 | 14.68 | 0.78 | 4.59|
|bench_string_upper[100000] | 125.52 | 20.99 | 0.83 | 5.98|
|bench_string_lower[1000000] | 67.39 | 15.11 | 0.78 | 4.46|
|bench_string_upper[1000000] | 125.58 | 20.58 | 0.84 | 6.10|